Process blocks without committing between transactions.
